Add workflow to update add-on releases
The workflow will update the add-on releases on repository dispatch.
Extract common code to avoid duplication between manual and automatic
task.
