+++ .github/workflows/update-add-on-release.yml
@@ -1,0 +1,23 @@
+name: Update Add-on Release
+
+on:
+  repository_dispatch:
+    types: 'add-on-release'
+
+jobs:
+  add-on-release:
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout
+      uses: actions/checkout@v2
+      with:
+        fetch-depth: 0
+    - name: Setup Java
+      uses: actions/setup-java@v1
+      with:
+        java-version: 8
+    - name: Update Add-on Release and Create Pull Request
+      run: echo "ADD_ON_DATA=${{ github.event.client_payload }}" >> $GITHUB_ENV
+      run: ./gradlew updateAndCreatePullRequestAddOnRelease --envVar=ADD_ON_DATA
+      env:
+        ZAPBOT_TOKEN: ${{ secrets.ZAPBOT_TOKEN }}
