+++ buildSrc/src/main/java/org/zaproxy/gradle/AddOnReleaseData.java
@@ -1,0 +1,81 @@
+/*
+ * Zed Attack Proxy (ZAP) and its related class files.
+ *
+ * ZAP is an HTTP/HTTPS proxy for assessing web application security.
+ *
+ * Copyright 2021 The ZAP Development Team
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.zaproxy.gradle;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.io.IOException;
+import java.util.List;
+
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+@JsonInclude(value = Include.NON_NULL)
+public class AddOnReleaseData {
+
+    @JsonProperty(value = "addons")
+    private List<Release> addOns;
+
+    public AddOnReleaseData() {}
+
+    public List<Release> getAddOns() {
+        return addOns;
+    }
+
+    public void setAddOns(List<Release> addOns) {
+        this.addOns = addOns;
+    }
+
+    public static AddOnReleaseData read(String file) {
+        try {
+            return new ObjectMapper().readValue(file, AddOnReleaseData.class);
+        } catch (IOException e) {
+            throw new TaskException("Failed to read the add-on release data: " + e.getMessage(), e);
+        }
+    }
+
+    @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+    @JsonInclude(value = Include.NON_NULL)
+    public static class Release {
+
+        private String url;
+        private String checksum;
+
+        public Release() {}
+
+        public String getUrl() {
+            return url;
+        }
+
+        public void setUrl(String url) {
+            this.url = url;
+        }
+
+        public String getChecksum() {
+            return checksum;
+        }
+
+        public void setChecksum(String checksum) {
+            this.checksum = checksum;
+        }
+    }
+}
